refactor: 예약 결제 API 결제 도메인 -> 예약 도메인으로 이동 및 엔드포인트 변경
