--- src/main/kotlin/kr/hhplus/be/server/api/payment/controller/dto/response/CreatePaymentResponse.kt
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.api.payment.controller.dto.response
-
-data class CreatePaymentResponse(
-    val id: String,
-)
