+++ src/main/kotlin/kr/hhplus/be/server/api/reservation/controller/dto/request/CreatePaymentRequest.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.api.payment.controller.dto.request
+package kr.hhplus.be.server.api.reservation.controller.dto.request
 
 import java.util.*
