+++ src/main/kotlin/kr/hhplus/be/server/api/reservation/controller/dto/response/CreatePaymentResponse.kt
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.api.reservation.controller.dto.response
+
+data class CreatePaymentResponse(
+    val id: String,
+)
