+++ src/test/java/kr/hhplus/be/server/api/reservation/application/ReservationFacadeIT.kt
@@ -2,18 +2,25 @@ package kr.hhplus.be.server.api.reservation.application
 
 import com.github.f4b6a3.tsid.TsidCreator
 import io.hhplus.cleanarchitecture.support.concurrent.ConcurrencyTestUtils
+import kr.hhplus.be.server.domain.balance.BalanceRepository
 import kr.hhplus.be.server.domain.concert.ConcertRepository
 import kr.hhplus.be.server.domain.concert.SeatRepository
 import kr.hhplus.be.server.domain.concert.model.CreateConcert
 import kr.hhplus.be.server.domain.concert.model.CreateSeat
+import kr.hhplus.be.server.domain.payment.PaymentRepository
 import kr.hhplus.be.server.domain.reservation.ReservationRepository
+import kr.hhplus.be.server.domain.reservation.model.CreateReservation
 import kr.hhplus.be.server.domain.reservation.model.Reservation
+import kr.hhplus.be.server.domain.token.TokenRepository
+import kr.hhplus.be.server.domain.token.model.CreateToken
 import kr.hhplus.be.server.support.IntegrationTestSupport
 import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import java.math.BigDecimal
+import java.util.*
 import java.util.concurrent.atomic.AtomicInteger
 
 @DisplayName("예약 Facade 통합 테스트")
@@ -22,6 +29,9 @@ class ReservationFacadeIT(
     private val reservationRepository: ReservationRepository,
     private val concertRepository: ConcertRepository,
     private val seatRepository: SeatRepository,
+    private val paymentRepository: PaymentRepository,
+    private val balanceRepository: BalanceRepository,
+    private val tokenRepository: TokenRepository,
 ) : IntegrationTestSupport() {
 
     @Nested
@@ -101,6 +111,69 @@ class ReservationFacadeIT(
             // then
             assertThat(successCount.get()).isEqualTo(1)
             assertThat(failCount.get()).isEqualTo(9)
+        }
+    }
+
+
+    @Nested
+    inner class `예약 결제` {
+        // given
+        val userId = TsidCreator.getTsid().toString()
+        val seatId = TsidCreator.getTsid().toString()
+        val payAmount = BigDecimal.valueOf(100)
+
+        val balance = balanceRepository.create(userId, BigDecimal(100))
+
+        val reservation = reservationRepository.save(
+            CreateReservation(
+                userId = userId,
+                seatId = seatId,
+                payAmount = payAmount,
+            )
+        )
+
+        val token = tokenRepository.save(
+            CreateToken(
+            userId = balance.userId,
+            token = UUID.randomUUID(),
+        )
+        )
+
+        // when
+        val result = reservationFacade.createPayment(reservation.id, token.token)
+
+        @Test
+        fun `성공 시 결제 내역이 남는다`() {
+            // when
+            val payment = paymentRepository.getById(result.paymentId)
+
+            // then
+            assertThat(payment).isNotNull
+        }
+
+        @Test
+        fun `성공 시 예약의 결제 금액만큼 잔고가 차감된다`() {
+            // when
+            val result = balanceRepository.getByUserId(balance.userId)
+
+            // then
+            assertThat(result.balance).isEqualTo(balance.balance.minus(reservation.payAmount))
+        }
+
+        @Test
+        fun `성공 시 대기열 토큰은 삭제된다`() {
+            // when
+            assertThatThrownBy { tokenRepository.getByToken(token.token) }
+                .isInstanceOf(Exception::class.java)
+        }
+
+        @Test
+        fun `성공 시 예약 상태는 COMPLETED 상태가 된다`() {
+            // when
+            val reservation = reservationRepository.getById(reservation.id)
+
+            // then
+            assertThat(reservation.status).isEqualTo(Reservation.Status.COMPLETED)
         }
     }
 }
